+++ internal/models/device_part/device_part.go
@@ -10,7 +10,7 @@ type DevicePart struct {
 	Id       int        `json:"id"`
 	DeviceId int        `json:"device_id"`
 	Price    uint64     `json:"price"`
-	Count    float32    `json:"count"`
+	Count    float64    `json:"count"`
 	Part     *part.Part `json:"part"`
 }
 
@@ -18,7 +18,7 @@ func NewEmptyDevicePart() *DevicePart {
 	return &DevicePart{}
 }
 
-func NewDevicePart(deviceId int, count float32, prt *part.Part) *DevicePart {
+func NewDevicePart(deviceId int, count float64, prt *part.Part) *DevicePart {
 	return &DevicePart{
 		DeviceId: deviceId,
 		Count:    count,
@@ -27,5 +27,5 @@ func NewDevicePart(deviceId int, count float32, prt *part.Part) *DevicePart {
 }
 
 func (dp *DevicePart) UpdatePrice() {
-	dp.Price = uint64(math.Ceil(float64(dp.Count * float32(dp.Part.Price))))
+	dp.Price = uint64(math.Ceil(float64(dp.Count) * float64(dp.Part.Price)))
 }
