+++ internal/models/project_device/project_device.go
@@ -9,7 +9,7 @@ import (
 type ProjectDevice struct {
 	Id        int            `json:"id"`
 	ProjectId int            `json:"project_id"`
-	Count     float32        `json:"count"`
+	Count     float64        `json:"count"`
 	Device    *device.Device `json:"device"`
 	Price     uint64         `json:"price"`
 }
@@ -18,7 +18,7 @@ func NewEmptyProjectDevice() *ProjectDevice {
 	return &ProjectDevice{}
 }
 
-func NewProjectDevice(prjId int, count float32, d *device.Device) *ProjectDevice {
+func NewProjectDevice(prjId int, count float64, d *device.Device) *ProjectDevice {
 	return &ProjectDevice{
 		ProjectId: prjId,
 		Count:     count,
@@ -27,5 +27,5 @@ func NewProjectDevice(prjId int, count float32, d *device.Device) *ProjectDevice
 }
 
 func (pd *ProjectDevice) UpdatePrice() {
-	pd.Price = uint64(math.Ceil(float64(pd.Count * float32(pd.Device.Price))))
+	pd.Price = uint64(math.Ceil(float64(pd.Count * float64(pd.Device.Price))))
 }
